+++ src/main/kotlin/io/github/darkstarworks/trialChamberPro/gui/GlobalSettingsView.kt
@@ -122,6 +122,7 @@ class GlobalSettingsView(
 
         gui.addPane(pane)
         gui.setOnGlobalClick { it.isCancelled = true }
+        gui.setOnGlobalDrag { it.isCancelled = true }
 
         return gui
     }