+++ src/main/kotlin/io/github/darkstarworks/trialChamberPro/gui/HelpMenuView.kt
@@ -63,6 +63,7 @@ class HelpMenuView(
 
         gui.addPane(pane)
         gui.setOnGlobalClick { it.isCancelled = true }
+        gui.setOnGlobalDrag { it.isCancelled = true }
 
         return gui
     }