+++ src/main/kotlin/io/github/darkstarworks/trialChamberPro/gui/LootTableListView.kt
@@ -83,6 +83,7 @@ class LootTableListView(
 
         gui.addPane(pane)
         gui.setOnGlobalClick { it.isCancelled = true }
+        gui.setOnGlobalDrag { it.isCancelled = true }
 
         return gui
     }