+++ src/main/kotlin/io/github/darkstarworks/trialChamberPro/gui/LootTypeSelectView.kt
@@ -82,6 +82,9 @@ class LootTypeSelectView(
         }, 0, 4)
 
         gui.addPane(pane)
+        gui.setOnGlobalClick { it.isCancelled = true }
+        gui.setOnGlobalDrag { it.isCancelled = true }
+
         return gui
     }