+++ src/main/kotlin/io/github/darkstarworks/trialChamberPro/gui/MainMenuView.kt
@@ -74,6 +74,7 @@ class MainMenuView(
 
         gui.addPane(pane)
         gui.setOnGlobalClick { it.isCancelled = true }
+        gui.setOnGlobalDrag { it.isCancelled = true }
 
         return gui
     }