+++ src/main/kotlin/io/github/darkstarworks/trialChamberPro/gui/SettingsMenuView.kt
@@ -70,6 +70,7 @@ class SettingsMenuView(
 
         gui.addPane(pane)
         gui.setOnGlobalClick { it.isCancelled = true }
+        gui.setOnGlobalDrag { it.isCancelled = true }
 
         return gui
     }